backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
opencv-python-headless
numpy
Pillow
supabase